+++ src/normalizations/normalize-editor.ts
@@ -9,7 +9,7 @@ import moment, { Moment } from 'moment';
 import { extractBody, extractHtmlBody } from '../commons/body-message-renderer';
 import { PREFS_DEFAULTS } from '../constants';
 import { CRB_XPARAMS, CRB_XPROPS } from '../constants/xprops';
-import { Editor, IdentityItem } from '../types/editor';
+import { Editor } from '../types/editor';
 import { Invite } from '../types/store/invite';
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
@@ -24,14 +24,6 @@ export const getVirtualRoom = (xprop: any): { label: string; link: string } | un
 
 	return undefined;
 };
-
-export const normaliseContact = (contact: { a: string; d: string }): IdentityItem => ({
-	...contact,
-	address: contact?.a,
-	fullName: contact?.d,
-	label: contact?.d ?? contact?.a,
-	value: contact?.a
-});
 
 const getAttendees = (attendees: any[], role: string): any[] =>
 	map(filter(attendees, ['role', role]), (at) =>
@@ -65,7 +57,7 @@ export type EventPropType = {
 	end: Date | Moment;
 };
 
-export const normalizeEditor = ({
+export const normalizeEditorWithoutOrganizer = ({
 	invite,
 	event,
 	id,
@@ -112,7 +104,6 @@ export const normalizeEditor = ({
 					inviteId: event?.resource?.inviteId,
 					reminder: invite?.alarmValue,
 					recur: invite.recurrenceRule,
-					organizer: invite?.organizer,
 					richText: extractHtmlBody(invite?.htmlDescription?.[0]?._content) ?? '',
 					plainText: extractBody(invite?.textDescription?.[0]?._content) ?? '',
 					uid: invite?.uid,