+++ src/shared/invite-response/invite-response.tsx
@@ -164,6 +164,11 @@ const InviteResponse: FC<InviteResponse> = ({
 					folders: calendarFolders,
 					isProposeNewTime: true,
 					panel: false,
+					organizer: {
+						label: `${requiredEvent.resource.organizer.d ?? requiredEvent.resource.organizer.a}`,
+						address: requiredEvent.resource.organizer.a,
+						fullName: requiredEvent.resource.organizer.d
+					},
 					disabled: {
 						title: true,
 						location: true,